Fix missing typeinfo in flex_bin
`flex` is used a prebuilt built tool in builds. ag/33676771 updates the
prebuilts, but the version of `flex` is missing the info in flexint.h
etc, likely due to the different paths of flex.skl and flexint.h inside
the sandbox. This CL should fix this by putting all the files in the
same location inside the sbox using `srcs`.

Test: Built out/host/linux-x86/bin/flex
Test: out/host/linux-x86/bin/flex -o out.cpp system/tools/aidl/aidl_language_l.ll
Test: grep typedef.*flex_int out.cpp # match found

Bug: b/393943813

Change-Id: If9741c5b359e104a438fe9bc6f713393eaae9b9c